// src/models/feedback.model.ts
export interface Feedback {
  id?: string;
  name?: string;
  phoneNumber?: string;
  isReviewed?: boolean;
  customerId?: string;
  email?: string;
  userType?: string;
  subject?: string;
  description?: string;
  archiveReason?: string;
  createdBy?: string;
  updatedBy?: string;
  createdAt?: string;
  updatedAt?: string;
  deletedAt?: string;
  deletedBy?: string;
}